refactor(contacts): extract MobileContactsPane (#166)
